+++ packages/forge-core/src/components/index.ts
@@ -1,6 +1,6 @@
 export { buildComponent } from './utils/buildComponent'
 export type { BuildComponentOptions } from './utils/buildComponent'
-export { component } from './component'
+export { component, isForgeComponent } from './component'
 export { coreComponents, HtmlBlock, CollectionBlock, TemplateWrapper, Fragment } from '../built-ins/components'
 export type { EvaluatedCollectionBlock } from '../built-ins/components'
 export type {